Using micrometers as a unit of distance, instead of arbitrary steps. Added capability to display axis settings

+++ application/System/Controllers/AxisMotionController/AxisMotionParameters.hpp
@@ -6,7 +6,7 @@ namespace ATC {
 struct AxisMotionParameters {
     uint8_t millimetersPerRotation;
     uint16_t motorStepsPerRotation;
-    uint16_t driverStepDivider;
+    uint8_t driverStepDivider;
     bool isClockwiseRotationForwardMovement;
     uint8_t stepErrorMargin;
 };
